pyproject.toml
[project]
name = "resonate-benchmark"
version = "0.1.0"
readme = "README.md"
authors = [
    { name = "Karn Wong", email = "karn@karnwong.me" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "opentelemetry-distro>=0.58b0",
    "opentelemetry-exporter-otlp>=1.37.0",
    "opentelemetry-propagator-b3>=1.37.0",
    "pydantic>=2.11.9",
    "resonate-sdk>=0.6.4",
    "uvicorn>=0.35.0",
]

[project.scripts]
foo = "resonate_benchmark.nodes.foo:main"
foo-otel = "resonate_benchmark.nodes.foo_otel:main"
main = "resonate_benchmark.main:main"

[build-system]
requires = ["uv_build>=0.8.2,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.18.1",
    "pip>=25.2",
    "pytest>=8.4.2",
]
